Latex feature pages: Add subheaders for spell level in the pdf
This patch adds a subsection for each spell level when the user
has selected to order spells by level.

The patch keeps count of the spell level and adds a new header
with every increase of spell level in the list of spells (i.e., headers
like: Cantrips; 1st Level Spells; 2nd Level Spells, and so on).

Add a test as well to check that the right spell headers also
show up for the spells sorted by level.