Latex feature pages: Enable DND-tables on fancy sheets
Dungeonsheets uses rst-to-latex to generate latex text.
For tables, rst-to-latex uses the latex supertabular
format. However, the RPGTeX latex style's DndTable
environment, as well as the DndLongTable environment
that we introduced, is not quite compatible with
supertabular. In particular, table headings result
in latex errors.

This patch modifies the rst-to-latex output so that it
generates a correct table header for the DndLongTable
environment. It also correctly sets the first row
for multipage tables.

The resulting tables are all slightly more compact
than the regular supertabular tables, and they look
better :-)

This patch only uses DndLongTable for the fancy sheets;
regular feature pages keep using the supertabular
output in order for them not to depend on the RPGTeX
latex style.