Latex feature pages: Sort feats by type
This patch gives the option to sort feats by type
instead of alphabetically. Basically, it just adds
the feats in the same order that dungeonsheets finds
them, which goes from character feat to class feats
to racial feats to background feats.

The patch implements this feature using a command-line
option. The user can specify --feats-by-type or -N to
order feats by type. Default is alphabetical order.