MSavage latex template: Switch to lualatex
The MSavage latex template was compiled using xelatex. To make this
work, an additional setting had to be used for dvipdfmx.cfg
(-dNOSAFER), which represents a (slight) security risk. Furthermore,
xelatex is very slow.

This patch switches to lualatex instead. It has different dependencies,
the most important one being the presence of luapstrics.lua, and the
character file requires the line \DocumentMetadata{} to correctly
deal with transparencies. As far as I can tell, by now luapstricks.lua
should be present on your average system with TexLive. It is on mine,
at least, on Slackware-15.0, which is moderately conservative in many
respects.