latex.py: Add local dnd.sty and dndntemplate.sty if installed
Dungeonsheets now packages the dnd latex template and the dnd
latex character sheet template packages as submodules. This
greatly eases using the latex code, but it also takes away
some opportunities for customisation.

This patch checks whether dnd.sty and dndtemplate.sty are
already installed in a local latex directory, and if so,
adds them to the texinputs path used in the latex subprocess
environment.

I also changed the order in which the various latex packages
are loaded in the MSavage latex format, so that, if
necessary, a user can pass options to the dnd package by
adding them to a locally checked out version of the
dndtemplate. And it would be possible at a later change
to switch the dndtemplate over to using tcolorbox with the
fitting library instead of the custom fitbox code used
now, in order to shrink font size and fit text within
specific boxes on the character sheet.