latex.py: Add local dnd.sty and dndntemplate.sty if installed
Dungeonsheets now packages the dnd latex template and the dnd
latex character sheet template packages as submodules. This
greatly eases using the latex code, but it also takes away
some opportunities for customisation.

This patch checks whether dnd.sty and dndtemplate.sty are
already installed in a local latex directory, and if so,
adds them to the texinputs path used in the latex subprocess
environment.