MSavage latex template: List all classes and levels
Previously, the MSavage latex template did not scale font
size in order to fit text inside specific spaces, which meant
that the levels and classes of a multiclass character would
overflow the text space for level and class.

With the current template, we can list all classes and levels.
Best tested with the sorcerer_ranger.py example.