Updates Dockerfile and test pipelines
This is a squashed commit of various changes by
Mat Savage to improve upon the github tests.

First, this patch updates the Docker file so that it is
compatible with lualatex. Furthermore, it adds a test
building all dungeonsheets examples in the docker workflow
file. And it refactors the python_ci workflow by only
testing all examples in python-3.12, to prevent the
test runs from taking so long. Finally, it pushes
the resulting docker image to the registry after tagging
it "latest".