fill_pdf_template: switch fallback from pdfrw to pypdf
Switch the fallback function to fill pdf forms to pypdf.
Pypdf has a simpler interface for form filling and appears
to be somewhat more robust that pdfrw. Also, this yields
pdfs in version 1.7 instead of 1.3, and it eliminates a
dependency on pdfrw. Finally, the pdfrw code appeared to
be buggy:

fill_pdf_template.py, line 454, in _make_pdf_pdfrw
    this_field = annot[FIELD][1:-1]
TypeError: 'NoneType' object is not subscriptable

Unfortunately, this patch is not able to flatten pdfs,
since this functionality isn't implemented in pypdf
(it isn't implemented in pdfrw either).