Latex feature pages: Switch fancy gm tables to DndLongTable
This patch changes the gm pages to use our internal
DndLongTable of the DndTable, instead of the one in the
RPGTeX dnd latex style.

This patch only uses DndLongTable for the fancy sheets;
regular feature pages keep using the supertabular
output in order for them not to depend on the RPGTeX
latex style.