Latex feature pages: Stretch DndLongTable to linewidth
In the RPGTeX dnd style, dnd tables are intended to be
exactly as wide as the text column. This patch takes
the column widths as calculated by docutils and
stretches all columns so that the resulting table
is exactly one column wide.

In most cases this just works, but I've noted a couple
of cases where tables end up being a little bit too
wide. Then again, you really need to look for this
to even see it.