Latex feature pages: Advanced feat sorting
This patch enables additional sorting for features
that are classes of other features. This works
as is for metamagic, in the sense that it sorts
all metamagic options under the metamagic feature,
by checking whether a metamagic option is the direct
child of the Metamagic class.

The same is intended for other feature options. The
intended relations are:
Bloodhunter:   Rites --> CrimsonRites
               BloodCurses --> BloodMaledict
               Formulas --> Formulas
Fighter:       Maneuver --> CombatSuperiority
               TrickShot --> AdeptMarksman
Warlock:       Invocation --> EldritchInvocation

However, these need some additional work to lay the
right class relations.