fillable_pdf_form.py: Use proficiencies_by_type for proficiencies
The fillable pdf form used proficiencies_text to display
the "other proficiencies and languages" information. However,
this resulted in a lot of duplication of weapon proficiencies.
Switch to proficiencies_by_type, which only displays weapon
and armor proficiencies as necessary. There are no changes
with regard to "other" proficiencies.

This patch makes it possible to integrate part of
proficencies_text into proficiencies_by_type.