Put all user option proficiencies in brackets
Various proficiency options exist that actually require
the player to choose something, such as:

three musical instruments of your choice; or
one set of artisan's tools.

These should not be listed as such on the character sheet.
Put each of these choices in brackets, like done for a
couple of such choices already, and for language options.
This way, we can filter these options out in character.py.