character.py: Fix adding subclasses
If no subclasses found, character.py will add "None"
as subclass for all classes if no subclasses are assigned.
However, this does not work for multiclass characters
where one class does and another does not yet have a
subclass, which will cause a failed assertion that
the number of subclasses is equal to the number of
classes.

Add None subclasses for the difference between
the number of classes and the number of sub
subclasses, based on the level.

This is a bugfix (no issue reported).