Features: Make FeatureSelector more robust
This patch makes parsing options for a FeatureSelector
more robust. First, it only compares selections with
the keys of the FeatureSelector options dictionary,
and second, it doesn't care whether developers have
defined options lowercase or uppercase.

In so doing, this fixes selecting a patron for the
bloodhunter Order of the Profane Soul:

The Bloodhunter Feature Selector for OtherwordlyPatron
lists the options for the patron choices capitalised.
However, the code in FeatureSelector expected them to
be lowercased. Make the FeatureSelector not care about
case.