latex.py: Silence a couple of docutils warnings
Docutils spews a truckload of FutureWarnings:

The default for the setting "use_latex_citations" will change to "True" in
Docutils 1.0.
The default for the setting "legacy_column_widths" will change to "False" in
Docutils 1.0.

This patch sets both use_latex_citations to True, which means that it might
be necessary in the future to set legacy_column_widths to False. However,
at present that causes various latex errors.