make_sheets.py: Don't set use_dnd_decorations for latex character sheet
This saves a lot of time compiling the latex character sheet.

Using the dnd latex style requires two passes of latex for correct
document compilation. So, if use_dnd_decorations is set (to select
the dnd latex style), latex.py executes two passes of latex.
However, this is not necessary for the latex character sheet.
Removing the use_dnd_decorations option reduces the compilation
time very much for the latex character sheet.