latex.py: More robust jinja latex filter without indentation
This patch removes indentation in latex by putting text inside
a group and setting \parindent to 0cm within that group.

This results in cleaner code, which would enable the user to
use, for instance, bulleted lists on their character sheet
also within the boxlatex code, which previously would have
crashed.