character: Add magic items to weight calculation
The weight calculation omitted the weight of magic items.

One would expect that this could be as simple as:
weight += sum([m.weight for m in self.magic_items]).
However, this doesn't work for magic items that are also
added to self.weapons, since these would be counted twice
in the weight calculation.