latex character template: Fix indentation the latex way
Some text fields in the latex character template need to be
formatted without indentation. This was previously done with
the "boxed" jinja filter, which basically replaced any pair
of newline characters with two slashes and one newline. This
however can break some rst-to-latex formatting, especially
when an unwary user tries to make a list in their character
file and by accident let rst-to-latex create an itemize
environment.

Incidentally, this patch makes the "boxed" jinja filter
obsolete.