equipment_reader: Be more forgiving on player input
The item_reader regex fails to match an empty string, which can be
caused be a player that accidentally leaves a comma at the end of
their equipment list. In this case, the weight_and_capacity_text
attribute will turn out not to be set, which causes jinja, and
the entire compilation to fail.