latex character template: Be more robust about hit_dice_current
The latex character sheet expected hit_dice_current to be a string
and would fail if it receives an int instead, while, logically,
the user might think that an int would be valid input.

Instead of a python string operation on hit_dice_current, use
the built-in replace filter of jinja, which does not seem to
care that its input is an int.