make_sheets: Move dealing with images after latex compilation
This patch deals with images after the try except loop for the
latex compilation has completed, and outside the "if not
use_tex_template" condition. This way, images are still
dealt with, even when lualatex is absent, and it makes
sure that all images are also processed with the latex
template, not only the portrait and symbol images.