spells / spells: Change Spell.__repr__ to "name"
The representation of spell was at one point changed to
{self.level} {self.name}, in a patch [1] changing the order
of spells by level and name. When that patch was reverted [2]
a couple of days later, but this revert did not include
Spell.__repr__. However, this means that the empty
character sheet jinja template would save spells in the
form of, e.g., "0 Light", the result of which in turn
fails compilation with makesheets.

1: commit 406b4d004d83873f10961d41ed9c70f4b917fb5b

2: commit 73e26b80d6a7807ef58e91f3a7c7ed252670714e